A few more changes - android

+++ ex_3/server/services/ChatServices.js
@@ -1,40 +1,35 @@
-const jwt = require("jsonwebtoken");
 const { Chat } = require("../models/Chat");
 const { User } = require("../models/User");
 const { Message } = require("../models/Message");
+const { decode, extractToken } = require("../tokens/JwtAuthenticator");
 
 // scheme: { id, users: ['tom', 'jerry']}
 // returns { id: string, user: User }
 exports.createChat = async (req, res) => {
     const { username: contactUsername } = req.body;
-    const token = req.headers.authorization.split(" ")[1];
-    const decodedToken = jwt.verify(token, "hello"); // Replace 'secret' with your actual secret
+    const token = extractToken(req);
+    const decodedToken = decode(token);
     const { username } = decodedToken;
     if (username === contactUsername) {
         return res
             .status(400)
-            .json({ message: "Cannot create chat with yourself." });
+            .json({ message: "Cannot create chat with yourself" });
     }
     try {
         //  Check if Users exist
         const currentUser = await User.findOne({ username });
         const contactUser = await User.findOne({ username: contactUsername });
         if (!currentUser || !contactUser) {
-            return res.status(404).json({ message: "User not found." });
-        }
-        //  Check if Chat exist
-        const existingChat = await Chat.findOne({
-            users: [username, contactUsername],
-        });
-        if (existingChat) {
-            return res.status(400).json({ message: "Chat already exists." });
+            return res.status(404).json({ message: "User not found" });
         }
 
-        const newChat = Chat.create({ users: [username, contactUsername] });
+        const newChat = await Chat.create({
+            users: [username, contactUsername],
+        });
         if (!newChat) {
             return res
                 .status(400)
-                .json({ message: "Error while creating new chat." });
+                .json({ message: "Error while creating new chat" });
         }
         res.status(201).json({ id: newChat.id, user: contactUser });
     } catch (error) {
@@ -43,8 +38,8 @@ exports.createChat = async (req, res) => {
 };
 
 exports.getContacts = async (req, res) => {
-    const token = req.headers.authorization.split(" ")[1];
-    const decodedToken = jwt.verify(token, "hello"); // Replace 'secret' with your actual secret
+    const token = extractToken(req);
+    const decodedToken = decode(token);
     const { username } = decodedToken;
 
     try {
@@ -55,11 +50,14 @@ exports.getContacts = async (req, res) => {
                     (user) => user !== username
                 );
                 const user = await User.findOne({ username: contactUser });
-                const lastMessage = await Message.findOne(
-                    { chatId: chat._id },
-                    { sender: 0 }
-                ).sort({ created: -1 });
-                return { id: chat._id, user, lastMessage };
+                let lastMessage = null;
+                const messages = await Message.find({ chatId: chat.id })
+                    .sort({ created: -1 })
+                    .limit(1);
+                if (messages.length > 0) {
+                    lastMessage = messages[0];
+                }
+                return { id: chat.id, user, lastMessage };
             })
         );
         res.status(200).json(finalChats);
@@ -70,21 +68,41 @@ exports.getContacts = async (req, res) => {
 
 exports.getChat = async (req, res) => {
     const chatId = req.params.id;
+    const token = extractToken(req);
+    const decodedToken = decode(token);
+    const { username } = decodedToken;
 
     try {
-        const chat = await Chat.findById(chatId)
-            .populate("users", "username displayName profilePic")
-            .populate({
-                path: "messages",
-                populate: {
-                    path: "sender",
-                    select: "username displayName profilePic",
-                },
-            });
-        if (!chat) {
-            return res.status(404).json({ message: "Chat not found." });
+        const chat = await Chat.findById(chatId);
+        if (!chat || !chat.users.includes(username)) {
+            return res.status(403).json({ message: "Unauthorized access" });
         }
-        res.status(200).json(chat);
+
+        const users = await User.find(
+            { username: { $in: chat.users } },
+            "username displayName profilePic"
+        );
+        const populatedChat = { id: chatId, users };
+
+        const messages = await Message.find({ chatId }, { chatId: 0 });
+        const populatedMessages = [];
+
+        for (const message of messages) {
+            const sender = users.find(
+                (user) => user.username === message.sender.username
+            );
+            const formattedMessage = {
+                id: message.id,
+                created: message.created,
+                sender,
+                content: message.content,
+            };
+            populatedMessages.push(formattedMessage);
+        }
+
+        populatedChat.messages = populatedMessages;
+
+        res.status(200).json(populatedChat);
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
@@ -93,15 +111,15 @@ exports.getChat = async (req, res) => {
 exports.createMessage = async (req, res) => {
     const chatId = req.params.id;
     const msg = req.body.msg;
-    const token = req.headers.authorization.split(" ")[1];
-    const decodedToken = jwt.verify(token, "hello"); // Replace 'secret' with your actual secret
+    const token = extractToken(req);
+    const decodedToken = decode(token);
     const { username } = decodedToken;
 
     try {
         // Check is chat exist
         const chat = await Chat.findById(chatId);
-        if (!chat) {
-            return res.status(404).json({ message: "Chat not found." });
+        if (!chat || !chat.users.includes(username)) {
+            return res.status(403).json({ message: "Unauthorized access" });
         }
         // Create new message
         const newMessage = await Message.create({
@@ -118,13 +136,20 @@ exports.createMessage = async (req, res) => {
 
 exports.getMessages = async (req, res) => {
     const chatId = req.params.id;
+    const token = extractToken(req);
+    const decodedToken = decode(token);
+    const { username } = decodedToken;
     try {
+        const chat = await Chat.findById(chatId);
+        if (!chat || !chat.users.includes(username)) {
+            return res.status(403).json({ message: "Unauthorized access" });
+        }
         const messages = await Message.find(
             { chatId },
             { id: "$_id", _id: 0, content: 1, created: 1, sender: 1 }
         ).sort({ created: -1 });
         if (!messages) {
-            return res.status(404).json({ message: "Messages not found." });
+            return res.status(404).json({ message: "Messages not found" });
         }
         res.status(200).json(messages);
     } catch (error) {
@@ -134,14 +159,25 @@ exports.getMessages = async (req, res) => {
 
 exports.deleteChat = async (req, res) => {
     const chatId = req.params.id;
+    const token = extractToken(req);
+    const decodedToken = decode(token);
+    const { username } = decodedToken;
 
     try {
-        // Delete the chat
-        const response = await Chat.deleteOne({ _id: chatId });
-        if (response.deletedCount === 0) {
-            return res.status(404).json({ message: "Chat not found." });
+        const chat = await Chat.findById(chatId);
+        if (!chat || !chat.users.includes(username)) {
+            return res.status(403).json({ message: "Unauthorized access" });
         }
-        res.status(200).json({ message: "Chat deleted." });
+        // Delete the chat
+        const deleteChatResult = await Chat.deleteOne({ _id: chatId });
+
+        if (deleteChatResult.deletedCount === 0) {
+            return res.status(404).json({ message: "Chat not found" });
+        }
+
+        // Delete the associated messages
+        await Message.deleteMany({ chatId });
+        res.status(200).json({ message: "Chat deleted" });
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
